requirements.txt
finance-datareader>=0.9.110
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
